Updated openstack/openstack Project: openstack/oslo.db  ed200283c17f5502a06b9081c1968c700219384a
Fix TestConnectionUtils to py3x compatibility

Tests in TestConnectionUtils class depends on database and DB connector,
so the simplest way to run them with python 3.X is to use same database
and connector for the both python versions. PostgreSQL and psycopg2 are
looks good for this.

Replaced ``mysql`` to ``postgresql`` in TestConnectionUtils.

Change-Id: Icb1d6324b58feed515e4eb885715cbf2195768cf