Updated openstack/openstack Project: openstack/trove  fb83a84e62489e822ae44bbef74faa467e443632
Obsolete oslo-incubator modules - exception

This change is part of a multi-part change set to handle obsolete and
graduated oslo modules. This change relates to the exception module
that is now obsolete. The change here is to delete the file from
openstack-common.conf and handle the change as it ripples into the
modules that used to include exception.

The change is modeled on the change that was implemented in
oslo-incubator when this module was deprecated. See review
https://review.openstack.org/#/c/39314

wsgi.py will be deprecated in another change so the only change being
made here is (effectively) to extensions.py

We can't delete the file in openstack/common because there is still a
dependency within oslo-incubator.

This change has been rebased on https://review.openstack.org/#/c/129714/

Blueprint: retire-unused-trove-modules
Partial-Bug: #1380789
Change-Id: I55e81693866cb15b3b06ee2ad1700f144023cc60