Updated openstack/openstack Project: openstack/tempest  10c0e4fa4219b31334687894075a5c04f42a4813
Skip test_update_router_admin_state for baremetal

The baremetal deployment relies on a flat network without tenant isolation.
The router servicing this single network should not be taken up and down,
especially if there are other tests running that may be accessing things
via public IP.  This is accounted for already during network/router/subnet
creation, but this test needs to be skipped as well.

Change-Id: I4064e134493085d90a0f0967b7b9ad640aadf75c
Closes-bug: #1400902