Updated openstack/openstack Project: openstack/tempest  03371c59ad7b1fcbd43d39f027af20c7d97ddedf
Make InvalidServiceTag inherit from TempestException

Exceptions which inherit from RestClientException should raise based
on an error response which is returned from REST API servers(nova-api,
etc). However, InvalidServiceTag is not now.
We have a plan RestClient is implemented in tempest-lib and related
exceptions also are moved to tempest-lib.
For doing that, this patch separates an unrelated exception from
RestClientException.

Change-Id: I49d50fb14b18c6fb263308a68d7ac6a97b1a9800