Updated openstack/openstack Project: openstack/nova  4c3b5fd1ec324595ffa64bd8fc961aab07cafc3d
Add backoff to ebtables retry

We need a backoff between ebtables retries. In some tempest tests we
have seen the retries complete in 100ms and still fail.

We now sleep for ebtables_retry_interval * loop count seconds. With
a default of 1.0 this means by default we sleep for 1.0s, 2.0s, and
3.0s before we finally giving up.

Change-Id: I0b9b664a592364bedd11124a1ec921d8ea011704
Partial-Bug: #1316621