Updated openstack/openstack Project: openstack/horizon  c70ec6f005afb13fecc9b3cce27af6fdf618ee2a
No delete volume action for volume with snapshots

Currently a volume with snapshots still has the "Delete Volume" menu
action.

This fix will not allow the "Delete Volume" menu action for a volume
that has snapshots. The main purpose of this change is to be consistent
with the existing behavior when the volume is not in a deletable state.

Closes-Bug: #1394015
Change-Id: If298e04952a29acbe3e8bd89e54065525ae5f30c