Updated openstack/openstack Project: openstack/ceilometer  bd5a8f4d49b690a1f90390b9af5c307290400a8e
Modify tests to support ordering of wsme types

When PYTHONHASHSEED is enabled the ordering of the wsme types are changed.
This fails the tests that take the type order into account when an assertEqual
is checked on the error message. The assertEqual is now changed to an
assertIn where the ordering of the types do not matter.

Change-Id: I0f445013802273478f5f448fa0350b9a1b140414
Closes-bug: #1399631