Updated openstack/openstack Project: openstack/glance  a502e4c8cc8584196cfb646f63f89c8510763058
Remove reliance on import order of oslo.db mods

This change fixes a couple of import statements that rely on some import
order and a bit of magic to make the oslo.db.options module visible
under the oslo.db package without importing it explicitly.

The old version of the imports relies on Python's import machinery to
attach sub-modules to packages after import (see
http://paste.openstack.org/show/148711/). It is safer to explicitly
import the module we need, to ensure that we are not relying on
side-effects of importing other modules to configure global state for
the application.

Change-Id: Ia2772030915061ff373e7b7dc77f161146b56e59