Updated openstack/openstack Project: openstack/nova  f77603c15bc4f580fec6571653463c9331b3819e
Extend clean_shutdown to the compute rpc layer

Adds the clean_shutdown parameter added to stop, resize, rescue,
shelve and shelve_offload in Juno to the compute rpc layer.  This
will in turn allow us to extend the API to provide a
forced-shutdown option to those operations.

Change-Id: Ie24c66e5006e64ef8b45d9d5056cc7876481a2f8
Partially-Implements: blueprint user-defined-shutdown