Updated openstack/openstack Project: openstack/nova  e1f3d037f3d886f080eb4a9ca26107eb35907cd5
Add support for clean_shutdown to stop in compute api layer

Change Iec3dfd17725440958aac395ebc471e51afd6522e added
clean_shutdown support to the compute manager in Juno.
This change is part of a set to extend that to the compute
API layer, which will in turn allow us to extend the REST API
to provide a forced-shutdown option to those operations.

Change-Id: I38707cf869705f983ab35e549d95dc4be4811ece
Partially-Implements: blueprint user-defined-shutdown