Updated openstack/openstack Project: openstack/nova  74f1415984e21949c58a2477bdf1589d083ab863
Add support for clean_shutdown to shelve in compute api layer

Change Iec3dfd17725440958aac395ebc471e51afd6522e added
clean_shutdown support to the compute manager in Juno.
This change is part of a set to extend that to the compute
API layer, which will in turn allow us to extend the REST API
to provide a forced-shutdown option to those operations.

Change-Id: I24dfef04cd0cddbff72267b93e0148b8fb204717
Partially-Implements: blueprint user-defined-shutdown