Updated openstack/openstack Project: openstack/nova  05af432e31e410f947417c45ede9f8c1927c659b
Add support for clean_shutdown to rescue in compute api layer

Change Iec3dfd17725440958aac395ebc471e51afd6522e added
clean_shutdown support to the compute manager in Juno.
This change is part of a set to extend that to the compute
API layer, which will in turn allow us to extend the REST API
to provide a forced-shutdown option to those operations.

Also adds a couple of missing unit test cases for rescue to the
compute_api tests

Change-Id: I5f00f63742fac89fea65149582f36866146e86f5
Partially-Implements: blueprint user-defined-shutdown