Updated openstack/openstack Project: openstack/python-keystoneclient  bfbc1a9bfe5c68e5e63ff780da6f4477ad484acf
Expose version matching functions to the public

The functions to match a version or convert a string version number into
a tuple have shown to be useful in at least auth_token middleware. I
think this is also better as _discover should really only be a shadow
for the discover file because of the circular dependency problems.
_discover shouldn't really need to be used even within client.

Closes-Bug: #1400998
Change-Id: Icf700c30d01e0700e437437a23e63a7f100ce4d3