Updated openstack/openstack Project: openstack/keystonemiddleware  791948cf07cca23b225915c32821dde3a10ce1e7
Use new ksc features in User Token Plugin

When the user token plugin was created some problems were noticed.

Firstly the way the auth_ref was being constructed meant that it would
try and retrieve the token from the body of the message which wouldn't
work for v3 or PKI tokens. To overcome this we stored the token data in
the plugin and added the ability to override the token data in the
auth_ref.

Secondly that there was no way to signal that this plugin couldn't be
reauthenticated and it would therefore retry unauthenticated requests
that we knew wouldn't work.

Both of these issues were addressed in keystoneclient and they should be
updated to work correctly in auth_token middleware.

Change-Id: Ib1c772e55f7c7b622d0c2a55d87d77d2dc30d4bb