Updated openstack/openstack Project: openstack/nova  1eaa53de5e6534833581ca2c8b5b85808c6f3a26
Replace stubs with mocks

This change updates the test cases for volume encryptors to use mocks
rather than stubs. The impetus for this change is comments on another
change (https://review.openstack.org/#/c/124791) that request new
tests to be written using mock. Mixing stubs and mocks would be
confusing so this change removes the existing stubs entirely.

Change-Id: I99ff78edcae5eaccabf4ea4caf04b8406ffa2f80