Updated openstack/openstack Project: openstack/python-cinderclient  9690cc37422558c9d21f2d331bee92bedae4bde4
Remove cinderclient/tests from coverage report

Currently coverage reports on coverage of the tests themselves. This
makes the percentage coverage higher than if they are excluded and
therefore distorts the reports values for the codes as a whole.

This commit updates the .coveragerc files to remove
cinderclient/tests/* from the coverage report.

Change-Id: If7f1765017f3f0d55975219b12ff7bf149146931