Updated openstack/openstack Project: openstack/horizon  8436f893c9074e0f67993118888a2916254a3590
Fixes the Order of "launch" buttons on jobs panel

On the data processing/jobs panel, the first "launch" button is "Launch
on New Cluster". This patch make the  "Launch on Existing Cluster" as
the first button. "Launch on New Cluster" is moved down on the list.

Also when "Launch On Existing Cluster" becomes the default option, the
"down arrow" to expand the action choices wraps down to the next line.
So added css to fix this issue.

Change-Id: I6f29e673f7ecf0316e855a3c0c0176d3c081491c
Closes-bug: #1390362