Updated openstack/openstack Project: openstack/python-novaclient  5f2ad683ce8d1933f4ca77d4327dff96b2aca347
Reject non existent mock assert calls

assert_called and assert_not_called are not asserting the state
of the mock object but considered as mocked calls so mock
will never raise exception but always executed successfully

This change patches the Mock class during unit test
to raise an exception if a function called on a mock object
that's name starts with 'assert' and does not one
of the supported Mock assert calls.

This change also fix the unit test to call only the supported
assert function on mock object.

Change-Id: I036587f355e42e362ac2b70fb6755cca81d30b75