Updated openstack/openstack Project: openstack/taskflow  97dd12e3e09edcbb16dbe94d46476ba67a20f45b
Retain the existence of a 'EngineBase' until 0.7 or later

For the recently renamed 'EngineBase' -> 'Engine' cleanup it
would be nice to retain the 'EngineBase' old named class for a
deprecation cycle. To enable this use the newly made deprecation
function that allows for creating inheritable classes that emit
deprecation warnings when constructed.

Change-Id: Ia67c924bc5896fbdc59bea25bf08fd87954905d0