Updated openstack/openstack Project: openstack/cinder  5259bd7f6010f17a129e6bbb1d8e486e93b33b03
Fix use of invalid variable in tgt exists check

I added a check for tgt already exists here:
https://review.openstack.org/#/c/138173/

But, that was wrong, the err value is never going to be
set because of the exception.  This should have inspected
the exception object and checked against it.

So this patch makes it do what it was supposed to do.

Change-Id: I8c184bf1684c8592a36f749cdcb5493f6b65e52b