Updated openstack/openstack Project: openstack-infra/devstack-gate  ac035803de984b727c6b1da63d4ea41386f638c9
Give a more specific error message when giving up in git_remote_update

git_remote_update is intermittently failing in the gate but we don't
have a good error message in the devstack logs for this specific case
due to the retry loop (we don't get 100% failure on a query for the
error messages in there today).

This change adds a specific error message when we fail the retry
attempts and exit so we can fingerprint on this with an elastic recheck
query.

Related-Bug: #1365046

Change-Id: Ibb3269a534d5b28b2d7d8a531e30a77ba6011f2d