Updated openstack/openstack Project: openstack/neutron  0f349d297c8e3dd77cd49af1795c72de03dcfce7
Have L3 agent catch the correct exception

L3 agent imports the processutils module to catch exceptions that
wouldn't ever be thrown because the underlying execute() being
called is the one from neutron.agent.linux.utils which raises a
RuntimeError on failure.

Also, processutils is now part of oslo.concurrency. So when we
actually start using it, we'll use it from there.

Closes-Bug: 1401626
Change-Id: I43874e1b63a0ba7b01415cafe0538f4343057066