Updated openstack/openstack Project: openstack/python-glanceclient  27f70bbd586da20943471bd26e3965e6fed26fe7
Remove keystoneclient mocks

The majority of the mocks on keystoneclient are not needed. Instead we
use the fixtures that keystoneclient provides to create well formed
responses and use requests_mock to stub out those values when accessed.

This is a closer to real usage pattern and the way that most client
projects are now handling keystoneclient auth.

Change-Id: I294aca1d11c9c5cfb4b82ad16476af4000efee1c