Updated openstack/openstack Project: openstack/heat  74152b301627668d99e1d6709532658671c1b340
Always update nested stacks

There's no reason to not issue an update to a nested stack, so don't bother
trying to decide if its inputs have changed (and especially don't replace
it because it's in a failed state). Overriding _needs_update() also allows
us to remove what was effectively an isinstance check from the previous
handling of failed states.

Change-Id: Ia48ebfcb6c1b3a904f01a83f8f34d8c7a13bb2be
Related-Bug: #1383618