Updated openstack/openstack Project: openstack/requirements  0ad261d28597356939ad9684fc5bcc50dbdec4af
pymemcache and sysv_ipc for tooz

Tooz currently does not enforce global requirements check and is
hence drifting. We need the pymemcache for the memcached driver
and we need sysv_ipc for the ipc driver in tooz.

Tooz uses pymemcache instead of the historical memcache Python bindings
because it has many more features including:

- Access to the "noreply" flag, which can significantly increase the
  speed of writes.
- Configurable timeouts for socket connect and send/recv calls.
- Flexible, simple approach to serialization and deserialization.
- Python 3 support
- Unit tests!

Change-Id: If934a117039b4101190aba98dfe33ae491027956