Updated openstack/openstack Project: openstack/neutron  5d3f221993dccb997b75a25c016acf8c80d1fd31
Replace non-ovs_lib calls of run_vsctl with libary functions

There is already an OVSBridge.set_controller() function, so replace
calls to run_vsctl(['set-controller', ...]) with calls to
OVSBridge.set_controller().

This is also the only non-ovs_lib call to run_vsctl. This change
will also make it easier to abstract out the ovsdb-related code in
the future.

Change-Id: Ia5a3e2c759d2479a2e7088a2ab7db1d4e17d1ae2