Updated openstack/openstack Project: openstack/python-novaclient  e422109eb3ea14acdb1cd11fd46cc353359d3b3e
novaclient sort parameters support

Adds sorting support to the 'nova list' command.

--sort <key>[:<direction>]

The --sort parameter is comma-separated and is used to specify
one or more sort keys and directions. The direction defaults to
'desc' for each sort key and the user can supply 'asc' to
override.

Partially implements: blueprint nova-pagination

Change-Id: I635e017c7f9ab61812333983bfecccd6fce8d394