Updated openstack/openstack Project: openstack/keystone  eb608eebf0cd7822a735d8e1885b632600d48286
Fix update role without name using LDAP

When the keystone server was configured to use the LDAP assignment
backend and a role was updated and no name was included in the update
dict the operation would fail with a 500 Internal Server Error.

With this fix, an update without a name works.

Change-Id: I234a36301a54ad4796cd421c42e85b15e9415902
Closes-Bug: #1241134